financegpt-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
openai>=1.0.0
psycopg2-binary>=2.9.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf>=3.0.0
python-multipart>=0.0.5
httpx
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
langchain_chroma
